Rails: relocate sinatra controllers
This will allow us to gradually port actions to ActionController.
Leaving the controller specs where they are, as they should largely work
the same for both.
