+++ .rubocop_todo.yml
@@ -1,19 +1,19 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config --auto-gen-only-exclude --exclude-limit 400`
-# on 2022-12-18 06:47:47 UTC using RuboCop version 1.40.0.
+# on 2022-12-18 18:04:01 UTC using RuboCop version 1.40.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 34
+# Offense count: 35
 # Configuration parameters: Include, IgnoredGems, OnlyFor.
 # Include: **/*.gemfile, **/Gemfile, **/gems.rb
 Bundler/GemComment:
   Exclude:
     - 'Gemfile'
 
-# Offense count: 32
+# Offense count: 33
 # Configuration parameters: EnforcedStyle, Include, AllowedGems.
 # SupportedStyles: required, forbidden
 # Include: **/*.gemfile, **/Gemfile, **/gems.rb
@@ -138,7 +138,7 @@ RSpec/AlignRightLetBrace:
     - 'spec/tasks/fetch_feeds_spec.rb'
     - 'spec/utils/feed_discovery_spec.rb'
 
-# Offense count: 6
+# Offense count: 7
 RSpec/AnyInstance:
   Exclude:
     - 'spec/controllers/debug_controller_spec.rb'
@@ -638,8 +638,8 @@ Rails/SaveBang:
     - 'app/commands/users/change_user_password.rb'
     - 'app/commands/users/complete_setup.rb'
     - 'app/commands/users/create_user.rb'
-    - 'app/controllers/first_run_controller.rb'
-    - 'app/controllers/stories_controller.rb'
+    - 'app/controllers/sinatra/first_run_controller.rb'
+    - 'app/controllers/sinatra/stories_controller.rb'
     - 'app/repositories/feed_repository.rb'
     - 'app/repositories/story_repository.rb'
     - 'app/repositories/user_repository.rb'
@@ -706,7 +706,7 @@ Rails/WhereNot:
 # MethodsAcceptingSymbol: inject, reduce
 Style/CollectionMethods:
   Exclude:
-    - 'app/controllers/stories_controller.rb'
+    - 'app/controllers/sinatra/stories_controller.rb'
     - 'app/fever_api/response.rb'
 
 # Offense count: 8
@@ -753,10 +753,10 @@ Style/FrozenStringLiteralComment:
     - 'app/commands/users/create_user.rb'
     - 'app/commands/users/sign_in_user.rb'
     - 'app/controllers/debug_controller.rb'
-    - 'app/controllers/feeds_controller.rb'
-    - 'app/controllers/first_run_controller.rb'
-    - 'app/controllers/sessions_controller.rb'
-    - 'app/controllers/stories_controller.rb'
+    - 'app/controllers/sinatra/feeds_controller.rb'
+    - 'app/controllers/sinatra/first_run_controller.rb'
+    - 'app/controllers/sinatra/sessions_controller.rb'
+    - 'app/controllers/sinatra/stories_controller.rb'
     - 'app/fever_api/authentication.rb'
     - 'app/fever_api/read_favicons.rb'
     - 'app/fever_api/read_feeds.rb'
@@ -933,7 +933,7 @@ Style/InlineComment:
   Exclude:
     - 'app/utils/opml_parser.rb'
 
-# Offense count: 694
+# Offense count: 695
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: IgnoreMacros, AllowedMethods, IgnoredMethods, AllowedPatterns, IgnoredPatterns, IncludedMacros, AllowParenthesesInMultilineCall, AllowParenthesesInChaining, AllowParenthesesInCamelCaseMethod, AllowParenthesesInStringInterpolation, EnforcedStyle.
 # SupportedStyles: require_parentheses, omit_parentheses
@@ -1081,7 +1081,7 @@ Style/PercentLiteralDelimiters:
 # SupportedStyles: same_as_string_literals, single_quotes, double_quotes
 Style/QuotedSymbols:
   Exclude:
-    - 'app/controllers/feeds_controller.rb'
+    - 'app/controllers/sinatra/feeds_controller.rb'
 
 # Offense count: 1
 # This cop supports safe autocorrection (--autocorrect).
@@ -1097,7 +1097,7 @@ Style/ReturnNil:
 # AllowedMethods: present?, blank?, presence, try, try!
 Style/SafeNavigation:
   Exclude:
-    - 'app/controllers/feeds_controller.rb'
+    - 'app/controllers/sinatra/feeds_controller.rb'
 
 # Offense count: 1
 # This cop supports safe autocorrection (--autocorrect).
@@ -1122,7 +1122,7 @@ Style/StaticClass:
     - 'app/utils/api_key.rb'
     - 'app/utils/content_sanitizer.rb'
 
-# Offense count: 18
+# Offense count: 20
 # This cop supports unsafe autocorrection (--autocorrect-all).
 Style/StringHashKeys:
   Exclude:
