+++ app/controllers/sinatra/first_run_controller.rb
@@ -1,9 +1,9 @@
-require_relative "../commands/feeds/import_from_opml"
-require_relative "../commands/users/create_user"
-require_relative "../commands/users/complete_setup"
-require_relative "../repositories/user_repository"
-require_relative "../repositories/story_repository"
-require_relative "../tasks/fetch_feeds"
+require_relative "../../commands/feeds/import_from_opml"
+require_relative "../../commands/users/create_user"
+require_relative "../../commands/users/complete_setup"
+require_relative "../../repositories/user_repository"
+require_relative "../../repositories/story_repository"
+require_relative "../../tasks/fetch_feeds"
 
 class Stringer < Sinatra::Base
   namespace "/setup" do
