+++ app/controllers/sinatra/stories_controller.rb
@@ -1,5 +1,5 @@
-require_relative "../repositories/story_repository"
-require_relative "../commands/stories/mark_all_as_read"
+require_relative "../../repositories/story_repository"
+require_relative "../../commands/stories/mark_all_as_read"
 
 class Stringer < Sinatra::Base
   get "/news" do
