+++ spec/controllers/feeds_controller_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-app_require "controllers/feeds_controller"
+app_require "controllers/sinatra/feeds_controller"
 
 describe "FeedsController" do
   let(:feeds) { [FeedFactory.build, FeedFactory.build] }
